refactor: extract Python shell into hyprstream-workers-python crate
Step 4 of the workers engine recut. The engine crate is now LANGUAGE-AGNOSTIC
(no Python).

Engine (hyprstream-workers-wasmtime):
- remove src/python.rs + src/python/mount.rs
- add generic public seam used by language profiles:
  * Sandbox::call_export_with_epoch (the epoch-bounded export driver,
    generalising the old python eval_with_epoch_deadline)
  * Sandbox::open_instance -> PersistentInstance { call_op(export,op,arg)
    -> (tag, payload) } — the generic packed-op ABI driver behind PyShell
  * TAG_OK made pub (shared ABI tag)
  * drop now-unused pub(crate) linker()/module() accessors
- tests/sandbox.rs: drive eval via call_export / call_export_with_epoch
  (security/DoS cases unchanged in intent); pyshell tests moved out

New crate (hyprstream-workers-python), mirrors hyprstream-tcl shape:
- lib.rs: PythonShell (over PersistentInstance) + PyResult + PyOp +
  eval()/eval_with_epoch_deadline() free fns + process_command + re-exports
- mount.rs: PythonMount(impl Mount) + PyCommand + create_mount_channel
  (tcl-style channel model: mount forwards PyCommand, owner drains)
- tests: mount E2E + persistent-shell vars/defs over the real guest
- added to workspace members

Verified (CI=1, real guest): engine 9 sandbox + 2 profile_b tests pass;
python 1 mount + 2 shell tests pass; clippy clean on both.
