+++ crates/hyprstream-workers-python/src/mount.rs
@@ -1,8 +1,8 @@
 //! `/lang/python/` VFS mount backed by the wasm-sandboxed python guest.
 //!
-//! Exposes a `/lang/python` layout driven by the [`PyShell`](crate::python::PyShell)
-//! over the wasm guest, so the interpreter is the sandboxed guest rather than a
-//! native one.
+//! Exposes a `/lang/python` layout driven by a [`PythonShell`](crate::PythonShell)
+//! over the wasm guest, so the interpreter is the sandboxed guest rather than a native
+//! one.
 //!
 //! Layout:
 //! - `/lang/python/eval`    — ctl: write a Python EXPRESSION, read `repr(result)`
@@ -13,65 +13,55 @@
 //!
 //! ## Why a command channel
 //!
-//! A [`PyShell`](crate::python::PyShell) owns a long-lived wasmtime `Store`. When the
-//! sandbox holds a VFS capability, its `vfs_*` host fns `blocking_send` on the proxy
-//! — so the shell MUST be driven from a NON-async (plain) thread, never a tokio
-//! worker. The async [`Mount`] methods therefore cannot call the shell directly:
-//! [`PythonMount`] holds a `tokio::sync::mpsc::Sender<PyCommand>` and forwards each
-//! request to a dedicated interpreter thread that owns the `PyShell` and replies
-//! over a oneshot, keeping the shell off the async runtime.
+//! A [`PythonShell`](crate::PythonShell) owns a long-lived wasmtime `Store`. When the
+//! sandbox holds a VFS capability, its `vfs_*` host fns `blocking_send` on the proxy —
+//! so the shell MUST be driven from a NON-async thread, never a tokio worker. The
+//! async [`Mount`] methods therefore cannot call the shell directly: [`PythonMount`]
+//! holds a `tokio::sync::mpsc::Sender<PyCommand>` and forwards each request to the
+//! thread that owns the shell, which replies over a oneshot.
 //!
-//! ## Daemon wiring
-//!
-//! Registering this mount into a running `/lang/python` namespace lives in the
-//! torch-bound `hyprstream` crate, not here. To plug in:
-//!
-//! ```ignore
-//! // In the daemon, where the per-session Namespace is built (alongside the Tcl
-//! // `/lang/tcl` mount):
-//! let sandbox = hyprstream_workers_wasmtime::Sandbox::from_bytes_for(GUEST_WASM, subject.clone())
-//!     .with_vfs(hyprstream_workers_wasmtime::vfs::VfsProxyHandle::new(
-//!         hyprstream_vfs::proxy::spawn_vfs_proxy(ns.clone(), subject.clone()),
-//!         subject.clone(),
-//!     ));
-//! let mount = hyprstream_workers_wasmtime::python::mount::PythonMount::spawn(sandbox, PER_CALL_FUEL);
-//! ns.mount("/lang/python", std::sync::Arc::new(mount))?;
-//! ```
-//!
-//! The mount owns the interpreter driver thread and joins it on `Drop`, so the
-//! daemon only needs to keep the `Arc<dyn Mount>` alive (in the Namespace).
+//! This mirrors `hyprstream-workers-tcl`'s `TclMount`/`TclCommand`/`create_mount_channel`
+//! model: build the channel with [`create_mount_channel`], mount
+//! `PythonMount::new(tx)` at `/lang/python`, and have the shell owner drain the
+//! `Receiver<PyCommand>` in its event loop (e.g. `ChatApp::tick()`), calling
+//! [`PythonShell::process_command`](crate::PythonShell::process_command).
 
 use async_trait::async_trait;
 use hyprstream_vfs::{DirEntry, Fid, Mount, MountError, Stat, Subject};
 use parking_lot::Mutex;
 
-use crate::python::{PyResult, PyShell};
-use crate::Sandbox;
+use crate::PyResult;
 
 // ─────────────────────────────────────────────────────────────────────────────
-// Commands sent from async Mount methods to the interpreter-owning thread.
+// Commands sent from async Mount methods to the shell-owning thread.
 // ─────────────────────────────────────────────────────────────────────────────
 
-/// A request to the thread that owns the [`PyShell`].
-enum PyCommand {
+/// A request from [`PythonMount`] to the thread owning the [`PythonShell`](crate::PythonShell).
+pub enum PyCommand {
+    /// Evaluate a Python expression. Response: the rendered [`PyResult`].
     Eval {
         code: String,
         resp: tokio::sync::oneshot::Sender<PyResult>,
     },
+    /// Execute Python statements. Response: captured stdout as a [`PyResult`].
     Exec {
         code: String,
         resp: tokio::sync::oneshot::Sender<PyResult>,
     },
+    /// List non-dunder global variable names.
     ListVars {
         resp: tokio::sync::oneshot::Sender<Vec<String>>,
     },
+    /// `repr` of one global variable (or `None` if absent).
     GetVar {
         name: String,
         resp: tokio::sync::oneshot::Sender<PyResult>,
     },
+    /// List callable global names.
     ListDefs {
         resp: tokio::sync::oneshot::Sender<Vec<String>>,
     },
+    /// `repr` of one callable global (or `None` if absent/not callable).
     GetDef {
         name: String,
         resp: tokio::sync::oneshot::Sender<PyResult>,
@@ -101,81 +91,33 @@ struct PyFid {
 }
 
 // ─────────────────────────────────────────────────────────────────────────────
-// The driver thread + mount.
+// PythonMount
 // ─────────────────────────────────────────────────────────────────────────────
 
-/// VFS mount exposing `/lang/python` over a wasm [`Sandbox`].
+/// VFS mount that proxies `/lang/python` requests to a [`PythonShell`](crate::PythonShell)
+/// over a command channel.
 ///
-/// `Send + Sync` (it only holds a channel sender + the driver `JoinHandle`); the
-/// `!`-async [`PyShell`] lives on the driver thread. Build with
-/// [`PythonMount::spawn`].
-///
-/// On `Drop`, the mount FIRST drops the command sender (closing the channel so the
-/// driver loop's `blocking_recv` returns `None` and the thread exits), THEN joins
-/// the thread. Doing both here — rather than in a separate guard — avoids the
-/// drop-ordering footgun where a separate guard would try to join while the sender
-/// is still alive (the channel never closes → join hangs forever).
+/// Mount at `/lang/python` in the namespace. The shell owner must drain the
+/// `Receiver<PyCommand>` returned by [`create_mount_channel`] in its event loop,
+/// forwarding each command to the shell via
+/// [`PythonShell::process_command`](crate::PythonShell::process_command).
 pub struct PythonMount {
-    /// `Option` so `Drop` can take (and thus drop) it BEFORE joining the thread.
-    tx: Option<tokio::sync::mpsc::Sender<PyCommand>>,
-    handle: Option<std::thread::JoinHandle<()>>,
-}
-
-impl Drop for PythonMount {
-    fn drop(&mut self) {
-        // Close the channel FIRST so the driver's blocking_recv returns None.
-        self.tx.take();
-        // Then join the driver thread (it will have exited its recv loop).
-        if let Some(h) = self.handle.take() {
-            let _ = h.join();
-        }
-    }
+    tx: tokio::sync::mpsc::Sender<PyCommand>,
 }
 
 impl PythonMount {
-    /// Spawn the interpreter driver thread over `sandbox` and return the mount.
-    /// `per_call_fuel` is the fuel budget per `eval`/`exec`.
-    ///
-    /// The driver thread opens the [`PyShell`] (a non-async context, so the
-    /// sandbox's `vfs_*` host fns may `blocking_send`) and serves `PyCommand`s until
-    /// the channel closes (i.e. until the mount is dropped). The mount owns the
-    /// thread and joins it on `Drop`.
-    pub fn spawn(sandbox: Sandbox, per_call_fuel: u64) -> Self {
-        let (tx, mut rx) = tokio::sync::mpsc::channel::<PyCommand>(16);
-        let handle = std::thread::Builder::new()
-            .name("hyprstream-workers-wasmtime-pyshell".into())
-            .spawn(move || {
-                // Open the persistent shell on THIS (plain) thread.
-                let mut shell: PyShell = match sandbox.open_shell(per_call_fuel) {
-                    Ok(s) => s,
-                    Err(e) => {
-                        eprintln!("pyshell: failed to open: {e}");
-                        return;
-                    }
-                };
-                // blocking_recv: this is a dedicated OS thread, not a tokio worker.
-                while let Some(cmd) = rx.blocking_recv() {
-                    serve(&mut shell, cmd);
-                }
-            })
-            .expect("spawn pyshell driver thread");
-        Self {
-            tx: Some(tx),
-            handle: Some(handle),
-        }
+    /// Create a new `PythonMount` from the sending half of a mount channel.
+    pub fn new(tx: tokio::sync::mpsc::Sender<PyCommand>) -> Self {
+        Self { tx }
     }
 
-    /// The command sender (always `Some` until `Drop`).
-    fn tx(&self) -> &tokio::sync::mpsc::Sender<PyCommand> {
-        self.tx.as_ref().expect("mount sender present until drop")
-    }
-
+    /// Send a command and await the shell response.
     async fn request<T>(
         &self,
         build: impl FnOnce(tokio::sync::oneshot::Sender<T>) -> PyCommand,
     ) -> Result<T, MountError> {
         let (resp_tx, resp_rx) = tokio::sync::oneshot::channel();
-        self.tx()
+        self.tx
             .send(build(resp_tx))
             .await
             .map_err(|_| MountError::Io("python interpreter gone".into()))?;
@@ -183,35 +125,6 @@ impl PythonMount {
             .await
             .map_err(|_| MountError::Io("python interpreter did not respond".into()))
     }
-}
-
-/// Serve one command against the shell (runs on the driver thread).
-fn serve(shell: &mut PyShell, cmd: PyCommand) {
-    match cmd {
-        PyCommand::Eval { code, resp } => {
-            let _ = resp.send(shell.eval(&code).unwrap_or_else(err_result));
-        }
-        PyCommand::Exec { code, resp } => {
-            let _ = resp.send(shell.exec(&code).unwrap_or_else(err_result));
-        }
-        PyCommand::ListVars { resp } => {
-            let _ = resp.send(shell.list_vars().unwrap_or_default());
-        }
-        PyCommand::GetVar { name, resp } => {
-            let _ = resp.send(shell.get_var(&name).unwrap_or_else(err_result));
-        }
-        PyCommand::ListDefs { resp } => {
-            let _ = resp.send(shell.list_defs().unwrap_or_default());
-        }
-        PyCommand::GetDef { name, resp } => {
-            let _ = resp.send(shell.get_def(&name).unwrap_or_else(err_result));
-        }
-    }
-}
-
-/// A trapped/host-side error surfaces to the guest as a Python-level error string.
-fn err_result(e: wasmtime::Error) -> PyResult {
-    PyResult::Err(format!("sandbox: {e}"))
 }
 
 /// Render a [`PyResult`] for the `eval`/`stdout` ctl read.
@@ -374,10 +287,22 @@ impl Mount for PythonMount {
     async fn clunk(&self, _fid: Fid, _caller: &Subject) {}
 }
 
+/// Create a `/lang/python` mount channel: the `Sender` goes to a
+/// [`PythonMount::new`], the `Receiver` is drained by the shell owner (which forwards
+/// each [`PyCommand`] to its [`PythonShell`](crate::PythonShell) via
+/// [`process_command`](crate::PythonShell::process_command)).
+pub fn create_mount_channel() -> (
+    tokio::sync::mpsc::Sender<PyCommand>,
+    tokio::sync::mpsc::Receiver<PyCommand>,
+) {
+    tokio::sync::mpsc::channel(16)
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::Sandbox;
+    use crate::PythonShell;
+    use hyprstream_workers_wasmtime::Sandbox;
     use std::path::PathBuf;
 
     fn guest_wasm() -> Option<Vec<u8>> {
@@ -413,9 +338,7 @@ mod tests {
                     "mount test: pyguest wasm not built but running under CI — build the \
                      pyguest and set HYPRSTREAM_PYGUEST_WASM"
                 );
-                eprintln!(
-                    "SKIP mount test: guest wasm not built (set CI=1 to make this a failure)"
-                );
+                eprintln!("SKIP mount test: guest wasm not built (set CI=1 to make this a failure)");
                 None
             }
         }
@@ -423,7 +346,8 @@ mod tests {
 
     /// End-to-end mount test against an in-memory VFS: write an expr to `eval`, read
     /// the result; exec sets a var, read it back under `vars/`; `readdir` lists the
-    /// layout, driven by the wasm guest.
+    /// layout, driven by the wasm guest. The shell runs on a dedicated thread that
+    /// drains the mount channel (mirroring the daemon owner loop).
     #[test]
     fn mount_eval_vars_defs_over_guest() {
         let Some(wasm) = guest_wasm_or_ci_fail() else {
@@ -436,8 +360,24 @@ mod tests {
             .unwrap();
 
         let subject = Subject::new("alice");
-        let sandbox = Sandbox::from_bytes_for(&wasm, subject.clone()).expect("load");
-        let mount = PythonMount::spawn(sandbox, 50_000_000_000);
+        let (tx, mut rx) = create_mount_channel();
+
+        // Shell owner thread: open the shell on a plain thread (vfs_* may
+        // blocking_send) and drain the mount channel.
+        let owner_wasm = wasm.clone();
+        let owner_subject = subject.clone();
+        let owner = std::thread::Builder::new()
+            .name("test-pyshell-owner".into())
+            .spawn(move || {
+                let sandbox = Sandbox::from_bytes_for(&owner_wasm, owner_subject).expect("load");
+                let mut shell = PythonShell::open(&sandbox, 50_000_000_000).expect("open shell");
+                while let Some(cmd) = rx.blocking_recv() {
+                    shell.process_command(cmd);
+                }
+            })
+            .expect("spawn owner");
+
+        let mount = PythonMount::new(tx);
 
         rt.block_on(async {
             // readdir root -> eval/stdout/vars/defs.
@@ -467,7 +407,7 @@ mod tests {
             let entries = mount.readdir(&vars_dir, &subject).await.unwrap();
             assert!(entries.iter().any(|e| e.name == "x"));
 
-            // Define a function; defs/ lists it, vars/ shows it too.
+            // Define a function; defs/ lists it.
             let stdout2 = mount.walk(&["stdout"], &subject).await.unwrap();
             mount
                 .write(&stdout2, 0, b"def f():\n    return 1\n", &subject)
@@ -479,5 +419,9 @@ mod tests {
 
             eprintln!("mount: eval/vars/defs over guest all green");
         });
+
+        // Close the channel so the owner thread exits, then join.
+        drop(mount);
+        owner.join().unwrap();
     }
 }
